#703 Moved specific methods to folders/files better reflecting their use, started refactoring long methods and deleting unused ones
