+++ cobigen-templates/templates-oasp4j/src/main/java/com/devonfw/cobigen/templates/oasp4j/utils/documentation/OpenApiDocumentationUtil.java
@@ -1,0 +1,8 @@
+package com.devonfw.cobigen.templates.oasp4j.utils.documentation;
+
+/**
+ * 
+ */
+public class OpenApiDocumentationUtil {
+
+}
