+++ cobigen-templates/templates-oasp4j/src/main/java/com/devonfw/cobigen/templates/oasp4j/utils/uml/Connector.java
@@ -1,4 +1,4 @@
-package com.devonfw.cobigen.templates.oasp4j.utils;
+package com.devonfw.cobigen.templates.oasp4j.utils.uml;
 
 /**
  * Connector is one association between classes in a class diagram. This class is used for storing that data, for later
